backend/src/invoke_ppt_generation/requirements.txt
boto3
botocore
logging
python-pptx
requests